move Home to src/app/time with own routes
