+++ README.md
@@ -7,25 +7,22 @@ Triggering the `serverTime()` server function or bumping `renderRealtimeClients(
 
 <img width="436" alt="Screenshot 2025-05-20 at 15 02 00" src="https://github.com/user-attachments/assets/4c3867bd-86f5-48e3-a43a-35de11cd2ea3" />
 
-### Home
+### Time
 ```tsx
-// src/app/pages/Home.tsx
+// src/app/time/Time.tsx
 import { Clock } from './Clock'
 import { ClientTimeButton } from './ClientTimeButton'
 import { ServerTimeButton } from './ServerTimeButton'
 import { ServerTime } from './ServerTime'
 import { BumpServerButton } from './BumpServerButton'
 
-export function Home() {
+export function Time() {
   console.log('Home RSC')
   return (
     <div className="flex flex-col items-center min-h-screen text-sm">
-      <h1 className="text-xl font-bold my-2">RedwoodSDK minimal RSC demo</h1>
-      <a
-        href="https://github.com/jldec/redwood-minimal-rsc#readme"
-        className="text-blue-600 p-2 underline mb-8 text-base"
-      >
-        github.com/jldec/redwood-minimal-rsc
+      <h1 className="text-xl font-bold my-2">Time</h1>
+      <a href="/" className="text-blue-600 p-2 underline mb-8 text-base">
+        Home
       </a>
       <Clock />
       <ClientTimeButton />
@@ -38,13 +35,46 @@ export function Home() {
 }
 ```
 
+### timeRoutes
+```ts
+// src/app/time/timeRoutes.ts
+import { realtimeRoute, renderRealtimeClients } from 'rwsdk/realtime/worker'
+import { render, route } from 'rwsdk/router'
+import { env } from 'cloudflare:workers'
+
+import { Document } from './Document'
+import { Time } from './Time'
+import { time } from './utils'
+
+async function handleBump() {
+  console.log('handle /api/bump')
+  await renderRealtimeClients({
+    durableObjectNamespace: env.REALTIME_DURABLE_OBJECT,
+    key: 'rwsdk-realtime-demo'
+  })
+  return new Response(time())
+}
+
+async function handleTime() {
+  console.log('handle /api/time')
+  return new Response(time())
+}
+
+export const timeRoutes = [
+  realtimeRoute(() => env.REALTIME_DURABLE_OBJECT),
+  render(Document, [route('/time', [Time])]),
+  route('/api/time', handleTime),
+  route('/api/bump', handleBump)
+]
+```
+
 ### Clock
 Live-updating clock rendered on the client.
 ```tsx
-// src/app/pages/Clock.tsx
+// src/app/time/Clock.tsx
 'use client'
 
-import { time } from '@/lib/utils'
+import { time } from './utils'
 import { useState, useEffect } from 'react'
 
 export function Clock() {
@@ -69,11 +99,11 @@ export function Clock() {
 ### ClientTimeButton
 Calls client-side function to display the time.
 ```tsx
-// src/app/pages/ClientTimeButton.tsx
+// src/app/time/ClientTimeButton.tsx
 'use client'
 
 import { useState } from 'react'
-import { time } from '@/lib/utils'
+import { time } from './utils'
 
 export function ClientTimeButton() {
   console.log('ClientTimeButton')
@@ -97,8 +127,8 @@ This is a RSC server function.
 
 NOTE: `'use server'` makes serverTime() callable from the client via fetch or realtime.
 ```ts
-// src/app/pages/serverTimeFunction.ts
-import { time } from '@/lib/utils'
+// src/app/time/serverTimeFunction.ts
+import { time } from './utils'
 
 export async function serverTime() {
   'use server'
@@ -110,7 +140,7 @@ export async function serverTime() {
 ### ServerTimeButton
 Calls server function, or fetches from `/api/time` if the `callFetch` prop is set.
 ```tsx
-// src/app/pages/ServerTimeButton.tsx
+// src/app/time/ServerTimeButton.tsx
 'use client'
 
 import { useState } from 'react'
@@ -148,8 +178,8 @@ export function ServerTimeButton({ callFetch = false }) {
 `ServerTime()` (with uppercase 'S') is a RSC (React Server Component) that displays the time. This is rendered as part of the initial page load, and re-rendered when serverTime() is called by the client.
 
 ```tsx
-// src/app/pages/ServerTime.tsx
-import { time } from '@/lib/utils'
+// src/app/time/ServerTime.tsx
+import { time } from './utils'
 
 export async function ServerTime() {
   console.log('ServerTime RSC')
@@ -179,36 +209,16 @@ initRealtimeClient({
 **worker.ts**
 ```ts
 // src/worker.ts
-import { Document } from '@/app/Document'
-import { Home } from '@/app/pages/Home'
-
 import { defineApp } from 'rwsdk/worker'
-import { index, render, route } from 'rwsdk/router'
-import { time } from '@/lib/utils'
-
-import { realtimeRoute, renderRealtimeClients } from 'rwsdk/realtime/worker'
-import { env } from 'cloudflare:workers'
+import { index, render } from 'rwsdk/router'
+import { timeRoutes } from './app/time/timeRoutes'
+import { Document } from './app/Document'
+import { Home } from './app/Home'
 export { RealtimeDurableObject } from 'rwsdk/realtime/durableObject'
 
-async function handleBump() {
-  console.log('handle /api/bump')
-  await renderRealtimeClients({
-    durableObjectNamespace: env.REALTIME_DURABLE_OBJECT,
-    key: 'rwsdk-realtime-demo'
-  })
-  return new Response(time())
-}
-
-async function handleTime() {
-  console.log('handle /api/time')
-  return new Response(time())
-}
-
 export default defineApp([
-  realtimeRoute(() => env.REALTIME_DURABLE_OBJECT),
-  render(Document, [index([Home])]),
-  route('/api/time', handleTime),
-  route('/api/bump', handleBump)
+  render(Document, [index(Home)]),
+  ...timeRoutes
 ])
 ```
 
@@ -216,7 +226,7 @@ export default defineApp([
 This button simulates a realtime update from the server by fetching /api/bump.
 
 ```tsx
-// src/app/pages/BumpServerButton.tsx
+// src/app/time/BumpServerButton.tsx
 'use client'
 
 import { useState } from 'react'
@@ -246,7 +256,7 @@ export function BumpServerButton() {
 ### time()
 Components share the time() function to get the formatted time.
 ```ts
-// src/lib/utils.ts
+// src/app/time/utils.ts
 export function formatTime(d: Date) {
   return new Intl.DateTimeFormat('en-US', {
     hour: '2-digit',
