+++ package.json
@@ -12,10 +12,10 @@
     "types": "wrangler types --include-runtime false"
   },
   "dependencies": {
-    "rwsdk": "0.0.84"
+    "rwsdk": "0.0.85"
   },
   "devDependencies": {
-    "@cloudflare/workers-types": "^4.20250521.0",
+    "@cloudflare/workers-types": "^4.20250522.0",
     "@tailwindcss/vite": "^4.1.7",
     "@types/node": "^22.15.21",
     "@types/react": "^19.1.5",
@@ -23,7 +23,7 @@
     "tailwindcss": "^4.1.7",
     "typescript": "^5.8.3",
     "vite": "^6.3.5",
-    "wrangler": "^4.16.0"
+    "wrangler": "^4.16.1"
   },
   "prettier": {
     "printWidth": 120,
