+++ src/app/time/Document.tsx
@@ -1,4 +1,4 @@
-import styles from "./styles.css?url";
+import styles from "../styles.css?url";
 
 export const Document: React.FC<{ children: React.ReactNode }> = ({
   children,
@@ -9,12 +9,13 @@ export const Document: React.FC<{ children: React.ReactNode }> = ({
     <head>
       <meta charSet="utf-8" />
       <meta name="viewport" content="width=device-width, initial-scale=1" />
-      <title>RedwoodSDK minimal RSC Demo</title>
+      <title>RedwoodSDK minimal RSC Time</title>
       <link rel="stylesheet" href={styles} />
       <link rel="icon" href="/favicon.jpg" />
     </head>
     <body>
       <div id="root">{children}</div>
+      <script>import("/src/client.tsx")</script>
     </body>
   </html>
 )};
