+++ src/app/time/ServerTime.tsx
@@ -1,4 +1,4 @@
-import { time } from '@/lib/utils'
+import { time } from './utils'
 
 export async function ServerTime() {
   console.log('ServerTime RSC')
