+++ src/app/time/utils.ts
